requirements.txt
absl-py==2.1.0
argbind==0.3.9
asttokens==2.4.1
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
docstring_parser==0.16
EbookLib==0.18
einops==0.8.0
executing==2.1.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
flatten-dict==0.4.2
fonttools==4.54.1
fsspec==2024.9.0
future==1.0.0
grpcio==1.66.2
gunicorn==23.0.0
html2text==2024.2.26
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
parler_tts @ git+https://github.com/huggingface/parler-tts.git@dcaed95e1cce6f616e3e1956f8d63f0f3f5dfe5f
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==3.19.6
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.4
PyPDF2==3.0.1
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
PyYAML==6.0.2
randomname==0.2.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.1
torch-stoi==0.2.3
torchaudio==2.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
